requirements.txt
google-cloud-aiplatform>=1.30.0
vertexai>=0.2.0
google-generativeai>=0.1.0
pytest>=7.0.0
Flask>=2.3.0
gunicorn>=21.2.0
python-dateutil==2.8.2
tinydb==4.7.0
python-dotenv==1.0.0
requests==2.31.0

# Optional / extra components (uncomment to enable local model support)
# transformers
# sentence-transformers
# fastapi
# uvicorn
# pydantic
# openai
